+++ src/prototypes/data-type-source.prototype.ts
@@ -72,10 +72,10 @@ export interface JSONAPIListResponse<Resource> {
   };
   readonly meta: {
     readonly page: {
-      readonly current_page: number;
+      readonly currentPage: number;
       readonly from: number;
-      readonly last_page: number;
-      readonly per_page: number;
+      readonly lastPage: number;
+      readonly perPage: number;
       readonly to: number;
       readonly total: number;
     };